[1.3.4] Merged revision(s) 4973, 5082, 5123, 5184, 5301, 5344, 5359, 5459, 5597, 5609, 5612-5616, 5619-5620, 5622, 5627, 5682, 5686, 5669, 5712-5714, 5718-5720, 5723-5725, 5727-5728 from trunk: r4973 (Fixed #6685 (request isAddEventHandler) #2 Patch by Kenix - Actually adds getEventHandlers) r5082 (Applied patch from sbx320: Fixed #4731 (onColShapeHit isn't triggered for towed vehicles server side)) r5123 (Fixed #7515 (Add "shared" export type in meta.xml)) r5184 (Fixed #5090 (GUI scrollpanes and scrollbars don't trigger onClientMouseEnter/Leave) Fixed #5870 ...)) r5301 (Prevented peds, created before onResourceStart has finished, from broadcasting packets) r5344 (Fixed CMatrix multiply) r5359 (Fixed #7570 (warpPedIntoVehicle after cancelEvent() of onVehicleStartEnter causes network trouble) #2) r5459 (Fixed #7615 (onPedWasted not triggered, when ped died because the vehicle he was in, exploded.)) r5597 (Refactored and renamed things relating to Lua files that are not cached on the client.) r5609 (Added Lua source encryption option) r5612 (Added playSFX ("Extended fix" of #1838 (Enable playMissionAudio and extend its functionality))...) r5613 (Update for r5612 - Added missing seeking (-> only the first sound worked), fixed a warning) r5614 (Update for r5612 - Fixed sound bank offset, replaced container id by enum) r5615 (Update for r5612 - Final offset fix) r5616 (Added some compiler asserts to interfaces from r5612) r5619 (Added SettingHUDMatchAspectRatio, SettingAspectRatio to dxGetStatus) r5620 (Added the ability to cancel onClientKey. Disabled for screenshot cancelling onClientKey ...) r5622 (Added escape to onClientKey (can only be cancelled once in a row)) r5627 (GTA SFX: Added 3D support ("playSFX3D") Game SFX volume will now be applied on new sfx ...) r5682 (Added encryption support to loadstring) r5686 (Fixed #7721 (cancelEvent() onClientKey escape key bug)) r5669 (Improved script performance slightly, and removed redundant code) r5712 (Fixed server createColPolygon) r5713 (Fixed #7757 (setAccountData clips the digits after the decimal point)) r5714 (Fixed a crash when destroying an object in onClientColShapeHit / onClientElementColShapeHit ...) r5718 (Refactored client settings to remove some duplicate code) r5719 (Test for #7768 (Vehicle door desync)) r5720 (Fixed some issues found by PVS Studio http://www.viva64.com) r5723 (Removed some debug code. Fixed some compile warnings. Fixed a crash.) r5724 (Updated sqlite to 3.7.17) r5725 (Fixed lightweight sync packet being misread on the client sometimes) r5727 (Fixed getLatentEventHandles sometimes returning false instead of an empty table) r5728 (Fixed more compile warnings.  Tweaked some packet priorities. Adjusted error level of some ...)

+++ branches/1.3.4/MTA10/mods/shared_logic/CClientStreamer.cpp
@@ -538,7 +538,7 @@ void CClientStreamer::Restream ( bool bMovedFar )
                 if ( bReachedLimit )
                 {
                     // Add to the list that might be streamed in during the final phase
-                    if ( ClosestStreamedOutList.size() < iMaxIn )     // (only add if there is a chance it will be used)
+                    if ( (int)ClosestStreamedOutList.size() < iMaxIn )     // (only add if there is a chance it will be used)
                         ClosestStreamedOutList.push_back( pElement );
                 }
                 else
